+++ Database/DaoObjects/Kvittering.cs
@@ -11,7 +11,7 @@ namespace Kvitteringer.Database.DaoObjects
         public int kvitId { get; set; }
         public DateOnly købsDato { get; set; }
         public DateOnly slutDato { get; set; }
-        public int ordreNummer { get; set; }
+        public long ordreNummer { get; set; }
         public string email { get; set; }
         public int firmaId { get; set; }
         public string firmaNavn { get; set; }
@@ -19,7 +19,7 @@ namespace Kvitteringer.Database.DaoObjects
         public double produktPris { get; set; }
 
 
-        public Kvittering(DateOnly købsDato, DateOnly slutDato, int ordreNummer, string email, string firmaNavn, string produktNavn, double produktPris)
+        public Kvittering(DateOnly købsDato, DateOnly slutDato, long ordreNummer, string email, string firmaNavn, string produktNavn, double produktPris)
         {
             this.købsDato = købsDato;
             this.slutDato = slutDato;
@@ -30,7 +30,7 @@ namespace Kvitteringer.Database.DaoObjects
             this.produktNavn = produktNavn;
             this.produktPris = produktPris;
         }
-        public Kvittering(int kvitId, DateOnly købsDatoOnly, DateOnly slutDato, int ordreNummer, int firmaId, string email, string produktNavn, double produktPris)
+        public Kvittering(int kvitId, DateOnly købsDatoOnly, DateOnly slutDato, long ordreNummer, int firmaId, string email, string produktNavn, double produktPris)
         {
             this.kvitId = kvitId;
             this.købsDato = købsDato;
@@ -43,7 +43,7 @@ namespace Kvitteringer.Database.DaoObjects
         }
 
       
-        public Kvittering(string firmaNavn, DateOnly købsDato, DateOnly slutDato, int ordreNummer, string email, string produktNavn, double produktPris)
+        public Kvittering(string firmaNavn, DateOnly købsDato, DateOnly slutDato, long ordreNummer, string email, string produktNavn, double produktPris)
         {
             this.firmaNavn = firmaNavn;
             this.købsDato = købsDato;
